projection: modified mean projection to use floating point values
Backport from a580a2a57f10132eacd172dcd8b4966f29cf2846 and 39ce31203089ca8591efcca3da56b48781da5825

+++ src/odemis/acq/stream/_projection.py
@@ -1647,7 +1647,7 @@ class SinglePointSpectrumProjection(DataProjection):
 
         mean = datasum / n
 
-        return model.DataArray(mean.astype(spec2d.dtype), md)
+        return model.DataArray(mean, md)
 
     def projectAsRaw(self):
         return self._computeSpec()
